requirements.txt
# Dev environment
pip
autopep8

# App
flask
oracledb
python-dotenv
